Consolidate actions that control search params into SearchParamActions, rename SearchActions to SearchRequestActions to be more clear.
