+++ client/src/error/ErrorContainer.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 import ErrorComponent from './ErrorComponent'
 import { push, goBack } from 'react-router-redux'
 import { clearErrors } from '../actions/ErrorActions'
-import { clearSearch } from '../actions/SearchActions'
+import { clearSearch } from '../actions/SearchParamActions'
 
 const mapStateToProps = (state) => {
   return {
