+++ client/src/reducers/domain/results.js
@@ -1,7 +1,6 @@
 import Immutable from 'seamless-immutable'
-import { SEARCH_COMPLETE, COUNT_HITS } from '../../actions/SearchActions'
-import { FETCHED_GRANULES, CLEAR_GRANULES } from '../../actions/SearchActions'
-import { FACETS_RECEIVED } from '../../search/facet/FacetActions'
+import { SEARCH_COMPLETE, COUNT_HITS } from '../../actions/SearchRequestActions'
+import { FETCHED_GRANULES, FACETS_RECEIVED, CLEAR_GRANULES } from '../../actions/SearchRequestActions'
 
 export const initialState = Immutable({
   collections: {},
