+++ client/src/search/map/MapContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import MapComponent from './MapComponent'
-import { newGeometry, removeGeometry } from './MapActions'
+import { newGeometry, removeGeometry } from '../../actions/SearchParamActions'
 
 const mapStateToProps = (state) => {
   const { geoJSON } = state.behavior.search
