+++ client/test/reducers/domain/resultsSpec.js
@@ -1,8 +1,7 @@
 import '../../specHelper'
 import Immutable from 'seamless-immutable'
 import { results, initialState } from '../../../src/reducers/domain/results'
-import { fetchedGranules, clearGranules } from '../../../src/actions/SearchActions'
-import { FACETS_RECEIVED } from '../../../src/search/facet/FacetActions'
+import { fetchedGranules, clearGranules, FACETS_RECEIVED } from '../../../src/actions/SearchRequestActions'
 
 describe('The results reducer', function () {
